opt/ui_loader/requirements.txt
streamlit>=1.36
pandas>=2.2
openpyxl>=3.1
SQLAlchemy>=2.0
psycopg2-binary>=2.9
python-dotenv>=1.0
tinydb>=4.8